feat: added banner hint
